[Ionut] [[BR]] - More work on CSM. Reverted back to VSM to allow easier blurring and filtering [[BR]] - Rewrote FrameBuffer classes to fix MSAA and memory leaks that occurred on resize (double creation of textures) [[BR]] -- Removed FBType. FrameBuffer types depend only on attachment types and if it's MSAA enabled or not [[BR]] - Rewrote RenderState classes to be API agnostic [[BR]] - Bug fixes in math classes [[BR]] - Changed "resolution change" logic [[BR]] - Simplified some options in config.xml [[BR]] - Improved per-light shadow options [[BR]]
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@260 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 9dc7bcffbdf10cd722e14273527a25a6dc97ae65 [formerly 11a3ac4b70ed945cb6e53a6063bd6f426d1933d0] [formerly a0d62c188491780e55b580ad0d9cd0c50ca164fb [formerly f20332904347d6dc28548ba3e5e5a17759b586ca]]
Former-commit-id: bed0fcdfc9d4d934d38fe04bb7aacf4d6b462c72 [formerly 84204d5422012860667e829631d33451e798bae8]
Former-commit-id: 49fd2186aa0762d9a9071560060d663a53a8b746

+++ Source Code/Libs/src/ReCast/DetourCrowd/Source/DetourCrowd.cpp
@@ -923,7 +923,7 @@ void dtCrowd::checkPathValidity(dtCrowdAgent** agents, const int nagents, const 
 		bool replan = false;
 
 		// First check that the current location is valid.
-		const int idx = getAgentIndex(ag);
+		const __int64 idx = getAgentIndex(ag);
 		float agentPos[3];
 		dtPolyRef agentRef = ag->corridor.getFirstPoly();
 		dtVcopy(agentPos, ag->npos);
@@ -1109,7 +1109,7 @@ void dtCrowd::update(const float dt, dtCrowdAgentDebugInfo* debug)
 		if (overOffmeshConnection(ag, triggerRadius))
 		{
 			// Prepare to off-mesh connection.
-			const int idx = ag - m_agents;
+			const __int64 idx = ag - m_agents;
 			dtCrowdAgentAnimation* anim = &m_agentAnims[idx];
 
 			// Adjust the path over the off-mesh connection.
